Library Reference

Fixed-Point Equations

SemiringFactorizations.sinv — Function

sinv(a)

Compute a quasi-inverse of a, i.e. an object a* satisfying

\[ a^* = 1 + a a^* = 1 + a^* a.\]

SemiringFactorizations.sldiv — Function

sldiv(a, b)

Solve the linear fixed-point equation

\[ ax + b = x,\]

where a, b, and x are elements of a semiring.

SemiringFactorizations.sldiv! — Function

sldiv!(A, B::AbstractVecOrMat)

Solve the linear fixed-point equation

\[ AX + B = X,\]

over-writing B with the solution.

SemiringFactorizations.mtsldiv! — Function

mtsldiv!(A::SparseSemiringLU, B::AbstractMatrix)

A multi-threaded version of sldiv!.

Warning

This feature is experimental and subject to change.

SemiringFactorizations.srdiv — Function

srdiv(b, a)

Solve the linear fixed-point equation

\[ xa + b = x,\]

where a, b, and x are elements of a semiring.

SemiringFactorizations.srdiv! — Function

srdiv!(B::AbstractVecOrMat, A)

Solve the linear fixed-point equation

\[ XA + B = X,\]

over-writing B with the solution.

SemiringFactorizations.mtsrdiv! — Function

mtsrdiv!(B::AbstractMatrix, A::SparseSemiringLU)

A multi-threaded version of srdiv!.

Warning

This feature is experimental and subject to change.

LU Factorizations

SemiringFactorizations.slu — Function

slu(A::AbstractMatrix)

Compute an LU factorization of a semiring- valued matrix A.

slu(A::SparseMatricCSC; alg=AMF(), snd=Maximal())

Compute an LU factorization of a sparse semiring-valued matrix A, optionally specifying an elimination algorithm alg and supernode type snd.

SemiringFactorizations.slu! — Function

slu!(A::AbstractMatrix)

Compute an LU factorization of a semiring- valued matrix A, over-writing A with the factors.

SemiringFactorizations.SemiringLU — Type

SemiringLU{T, M <: AbstractMatrix{T}} <: AbstractSemiringLU{T}

An LU factorization of a semiring-valued matrix.

SemiringFactorizations.SparseSemiringLU — Type

SparseSemiringLU{T, I} <: AbstractSemiringLU{T}

An LU factorization of a sparse semiring- valued matrix.

SemiringFactorizations.SymbolicSemiringLU — Type

SymbolicSemiringLU{I}

A symbolic factorization of a sparse semiring-valued matrix.